Fix for Matlab 2016

+++ lib/gco-v3.0/src/gco_matlab.cpp
@@ -5,8 +5,8 @@
 #include "GCoptimization.h"
 
 #if !defined(MX_API_VER) || MX_API_VER < 0x07030000
-typedef int mwSize;
-typedef int mwIndex;
+typedef size_t mwSize;
+typedef size_t mwIndex;
 #endif
 
 namespace matrix{ namespace detail{ namespace noninlined{ namespace mx_array_api{
